Refactor GlobalParams. Move to core, and set up OpenAIDefaults as user interface.
