+++ cognitive/src/main/scala/com/microsoft/azure/synapse/ml/services/CognitiveServiceBase.scala
@@ -10,7 +10,7 @@ import com.microsoft.azure.synapse.ml.fabric.FabricClient
 import com.microsoft.azure.synapse.ml.io.http._
 import com.microsoft.azure.synapse.ml.logging.SynapseMLLogging
 import com.microsoft.azure.synapse.ml.logging.common.PlatformDetails
-import com.microsoft.azure.synapse.ml.param.ServiceParam
+import com.microsoft.azure.synapse.ml.param.{GlobalParams, ServiceParam}
 import com.microsoft.azure.synapse.ml.stages.{DropColumns, Lambda}
 import org.apache.http.NameValuePair
 import org.apache.http.client.methods.{HttpEntityEnclosingRequestBase, HttpPost, HttpRequestBase}
@@ -547,6 +547,21 @@ abstract class CognitiveServicesBaseNoHandler(val uid: String) extends Transform
   }
 
   override def transform(dataset: Dataset[_]): DataFrame = {
+    // check for empty params. Fill em with OpenAIDefaults
+    this.params
+      .filter(p => !this.isSet(p) && !this.hasDefault(p))
+      .foreach { p =>
+        GlobalParams.getParam(p) match {
+          case Some(v) =>
+            p match {
+              case serviceParam: ServiceParam[_] =>
+                setScalarParam(serviceParam.asInstanceOf[ServiceParam[Any]], v)
+              case param: Param[_] =>
+                set(param.asInstanceOf[Param[Any]], v)
+            }
+          case None =>
+        }
+      }
     logTransform[DataFrame](getInternalTransformer(dataset.schema).transform(dataset), dataset.columns.length
     )
   }
