+++ cognitive/src/main/scala/com/microsoft/azure/synapse/ml/services/openai/OpenAIChatCompletion.scala
@@ -34,7 +34,7 @@ class OpenAIChatCompletion(override val uid: String) extends OpenAIServicesBase(
   }
 
   override protected def prepareUrlRoot: Row => String = { row =>
-    s"${getUrl}openai/deployments/${getGlobalServiceParamValue(row, deploymentName)}/chat/completions"
+    s"${getUrl}openai/deployments/${getValue(row, deploymentName)}/chat/completions"
   }
 
   override protected[openai] def prepareEntity: Row => Option[AbstractHttpEntity] = {
