+++ cognitive/src/main/scala/com/microsoft/azure/synapse/ml/services/openai/OpenAIDefaults.scala
@@ -1,0 +1,9 @@
+package com.microsoft.azure.synapse.ml.services.openai
+
+import com.microsoft.azure.synapse.ml.param.GlobalParams
+
+object OpenAIDefaults {
+  def setDeploymentName(v: String): Unit = {
+    GlobalParams.setGlobalParam(OpenAIDeploymentNameKey, v)
+  }
+}
