+++ cognitive/src/test/scala/com/microsoft/azure/synapse/ml/services/openai/OpenAIDefaultsSuite.scala
@@ -1,0 +1,56 @@
+package com.microsoft.azure.synapse.ml.services.openai
+
+import com.microsoft.azure.synapse.ml.core.test.base.Flaky
+import org.apache.spark.sql.{DataFrame, Row}
+
+class OpenAIDefaultsSuite extends Flaky with OpenAIAPIKey {
+
+  import spark.implicits._
+
+  OpenAIDefaults.setDeploymentName(deploymentName)
+
+  def promptCompletion: OpenAICompletion = new OpenAICompletion()
+    .setCustomServiceName(openAIServiceName)
+    .setMaxTokens(200)
+    .setOutputCol("out")
+    .setSubscriptionKey(openAIAPIKey)
+    .setPromptCol("prompt")
+
+  lazy val promptDF: DataFrame = Seq(
+    "Once upon a time",
+    "Best programming language award goes to",
+    "SynapseML is "
+  ).toDF("prompt")
+
+  test("Completion w Globals") {
+    val fromRow = CompletionResponse.makeFromRowConverter
+    promptCompletion.transform(promptDF).collect().foreach(r =>
+      fromRow(r.getAs[Row]("out")).choices.foreach(c =>
+        assert(c.text.length > 10)))
+  }
+
+  lazy val prompt: OpenAIPrompt = new OpenAIPrompt()
+    .setSubscriptionKey(openAIAPIKey)
+    .setCustomServiceName(openAIServiceName)
+    .setOutputCol("outParsed")
+    .setTemperature(0)
+
+  lazy val df: DataFrame = Seq(
+    ("apple", "fruits"),
+    ("mercedes", "cars"),
+    ("cake", "dishes"),
+    (null, "none") //scalastyle:ignore null
+  ).toDF("text", "category")
+
+  test("OpenAIPrompt w Globals") {
+    val nonNullCount = prompt
+      .setPromptTemplate("here is a comma separated list of 5 {category}: {text}, ")
+      .setPostProcessing("csv")
+      .transform(df)
+      .select("outParsed")
+      .collect()
+      .count(r => Option(r.getSeq[String](0)).isDefined)
+
+    assert(nonNullCount == 3)
+  }
+}
